PW-2289: move details classes into details package, removed default subtype value for PayPalDetails
