+++ src/main/java/com/adyen/model/checkout/details/BilldeskWalletDetails.java
@@ -19,48 +19,49 @@
  * See the LICENSE file for more info.
  */
 
-package com.adyen.model.checkout;
+package com.adyen.model.checkout.details;
 
+import com.adyen.model.checkout.PaymentMethodDetails;
 import com.google.gson.annotations.SerializedName;
 
 import java.util.Objects;
 
 /**
- * LianLianPayDetails
+ * BilldeskWalletDetails
  */
 
-public class LianLianPayDetails implements PaymentMethodDetails {
-    @SerializedName("telephoneNumber")
-    private String telephoneNumber = null;
+public class BilldeskWalletDetails implements PaymentMethodDetails {
+    @SerializedName("issuer")
+    private String issuer = null;
 
     @SerializedName("type")
-    private String type = null;
+    private String type = "billdesk_wallet";
 
-    public LianLianPayDetails telephoneNumber(String telephoneNumber) {
-        this.telephoneNumber = telephoneNumber;
+    public BilldeskWalletDetails issuer(String issuer) {
+        this.issuer = issuer;
         return this;
     }
 
     /**
-     * Get telephoneNumber
+     * The issuer id of the shopper&#x27;s selected bank.
      *
-     * @return telephoneNumber
+     * @return issuer
      **/
-    public String getTelephoneNumber() {
-        return telephoneNumber;
+    public String getIssuer() {
+        return issuer;
     }
 
-    public void setTelephoneNumber(String telephoneNumber) {
-        this.telephoneNumber = telephoneNumber;
+    public void setIssuer(String issuer) {
+        this.issuer = issuer;
     }
 
-    public LianLianPayDetails type(String type) {
+    public BilldeskWalletDetails type(String type) {
         this.type = type;
         return this;
     }
 
     /**
-     * Get type
+     * **billdesk_wallet**
      *
      * @return type
      **/
@@ -81,23 +82,23 @@ public class LianLianPayDetails implements PaymentMethodDetails {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        LianLianPayDetails lianLianPayDetails = (LianLianPayDetails) o;
-        return Objects.equals(this.telephoneNumber, lianLianPayDetails.telephoneNumber) &&
-                Objects.equals(this.type, lianLianPayDetails.type);
+        BilldeskWalletDetails billdeskWalletDetails = (BilldeskWalletDetails) o;
+        return Objects.equals(this.issuer, billdeskWalletDetails.issuer) &&
+                Objects.equals(this.type, billdeskWalletDetails.type);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(telephoneNumber, type);
+        return Objects.hash(issuer, type);
     }
 
 
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append("class LianLianPayDetails {\n");
+        sb.append("class BilldeskWalletDetails {\n");
 
-        sb.append("    telephoneNumber: ").append(toIndentedString(telephoneNumber)).append("\n");
+        sb.append("    issuer: ").append(toIndentedString(issuer)).append("\n");
         sb.append("    type: ").append(toIndentedString(type)).append("\n");
         sb.append("}");
         return sb.toString();
