+++ src/main/java/com/adyen/model/checkout/details/DragonpayDetails.java
@@ -19,27 +19,70 @@
  * See the LICENSE file for more info.
  */
 
-package com.adyen.model.checkout;
+package com.adyen.model.checkout.details;
 
+import com.adyen.model.checkout.PaymentMethodDetails;
 import com.google.gson.annotations.SerializedName;
 
 import java.util.Objects;
 
 /**
- * GiropayDetails
+ * DragonpayDetails
  */
 
-public class GiropayDetails implements PaymentMethodDetails {
-    @SerializedName("type")
-    private String type = "giropay";
+public class DragonpayDetails implements PaymentMethodDetails {
+    @SerializedName("issuer")
+    private String issuer = null;
 
-    public GiropayDetails type(String type) {
+    @SerializedName("shopperEmail")
+    private String shopperEmail = null;
+
+    @SerializedName("type")
+    private String type = null;
+
+    public DragonpayDetails issuer(String issuer) {
+        this.issuer = issuer;
+        return this;
+    }
+
+    /**
+     * The Dragonpay issuer value of the shopper&#x27;s selected bank. Set this to an **id** of a Dragonpay issuer to preselect it.
+     *
+     * @return issuer
+     **/
+    public String getIssuer() {
+        return issuer;
+    }
+
+    public void setIssuer(String issuer) {
+        this.issuer = issuer;
+    }
+
+    public DragonpayDetails shopperEmail(String shopperEmail) {
+        this.shopperEmail = shopperEmail;
+        return this;
+    }
+
+    /**
+     * The shopper’s email address.
+     *
+     * @return shopperEmail
+     **/
+    public String getShopperEmail() {
+        return shopperEmail;
+    }
+
+    public void setShopperEmail(String shopperEmail) {
+        this.shopperEmail = shopperEmail;
+    }
+
+    public DragonpayDetails type(String type) {
         this.type = type;
         return this;
     }
 
     /**
-     * **giropay**
+     * Get type
      *
      * @return type
      **/
@@ -60,21 +103,25 @@ public class GiropayDetails implements PaymentMethodDetails {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        GiropayDetails giropayDetails = (GiropayDetails) o;
-        return Objects.equals(this.type, giropayDetails.type);
+        DragonpayDetails dragonpayDetails = (DragonpayDetails) o;
+        return Objects.equals(this.issuer, dragonpayDetails.issuer) &&
+                Objects.equals(this.shopperEmail, dragonpayDetails.shopperEmail) &&
+                Objects.equals(this.type, dragonpayDetails.type);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(type);
+        return Objects.hash(issuer, shopperEmail, type);
     }
 
 
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append("class GiropayDetails {\n");
+        sb.append("class DragonpayDetails {\n");
 
+        sb.append("    issuer: ").append(toIndentedString(issuer)).append("\n");
+        sb.append("    shopperEmail: ").append(toIndentedString(shopperEmail)).append("\n");
         sb.append("    type: ").append(toIndentedString(type)).append("\n");
         sb.append("}");
         return sb.toString();
