+++ src/main/java/com/adyen/model/checkout/details/EntercashDetails.java
@@ -19,69 +19,49 @@
  * See the LICENSE file for more info.
  */
 
-package com.adyen.model.checkout;
+package com.adyen.model.checkout.details;
 
+import com.adyen.model.checkout.PaymentMethodDetails;
 import com.google.gson.annotations.SerializedName;
 
 import java.util.Objects;
 
 /**
- * SepaDirectDebitDetails
+ * EntercashDetails
  */
 
-public class SepaDirectDebitDetails implements PaymentMethodDetails {
-    @SerializedName("iban")
-    private String iban = null;
-
-    @SerializedName("ownerName")
-    private String ownerName = null;
+public class EntercashDetails implements PaymentMethodDetails {
+    @SerializedName("issuer")
+    private String issuer = null;
 
     @SerializedName("type")
-    private String type = "sepadirectdebit";
+    private String type = "entercash";
 
-    public SepaDirectDebitDetails iban(String iban) {
-        this.iban = iban;
+    public EntercashDetails issuer(String issuer) {
+        this.issuer = issuer;
         return this;
     }
 
     /**
-     * The International Bank Account Number (IBAN).
+     * The issuer id of the shopper&#x27;s selected bank.
      *
-     * @return iban
+     * @return issuer
      **/
-    public String getIban() {
-        return iban;
+    public String getIssuer() {
+        return issuer;
     }
 
-    public void setIban(String iban) {
-        this.iban = iban;
+    public void setIssuer(String issuer) {
+        this.issuer = issuer;
     }
 
-    public SepaDirectDebitDetails ownerName(String ownerName) {
-        this.ownerName = ownerName;
-        return this;
-    }
-
-    /**
-     * The name of the bank account holder.
-     *
-     * @return ownerName
-     **/
-    public String getOwnerName() {
-        return ownerName;
-    }
-
-    public void setOwnerName(String ownerName) {
-        this.ownerName = ownerName;
-    }
-
-    public SepaDirectDebitDetails type(String type) {
+    public EntercashDetails type(String type) {
         this.type = type;
         return this;
     }
 
     /**
-     * **sepadirectdebit**
+     * **entercash**
      *
      * @return type
      **/
@@ -102,25 +82,23 @@ public class SepaDirectDebitDetails implements PaymentMethodDetails {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        SepaDirectDebitDetails sepaDirectDebitDetails = (SepaDirectDebitDetails) o;
-        return Objects.equals(this.iban, sepaDirectDebitDetails.iban) &&
-                Objects.equals(this.ownerName, sepaDirectDebitDetails.ownerName) &&
-                Objects.equals(this.type, sepaDirectDebitDetails.type);
+        EntercashDetails entercashDetails = (EntercashDetails) o;
+        return Objects.equals(this.issuer, entercashDetails.issuer) &&
+                Objects.equals(this.type, entercashDetails.type);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(iban, ownerName, type);
+        return Objects.hash(issuer, type);
     }
 
 
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append("class SepaDirectDebitDetails {\n");
+        sb.append("class EntercashDetails {\n");
 
-        sb.append("    iban: ").append(toIndentedString(iban)).append("\n");
-        sb.append("    ownerName: ").append(toIndentedString(ownerName)).append("\n");
+        sb.append("    issuer: ").append(toIndentedString(issuer)).append("\n");
         sb.append("    type: ").append(toIndentedString(type)).append("\n");
         sb.append("}");
         return sb.toString();
