+++ src/main/java/com/adyen/model/checkout/details/MbwayDetails.java
@@ -19,48 +19,70 @@
  * See the LICENSE file for more info.
  */
 
-package com.adyen.model.checkout;
+package com.adyen.model.checkout.details;
 
+import com.adyen.model.checkout.PaymentMethodDetails;
 import com.google.gson.annotations.SerializedName;
 
 import java.util.Objects;
 
 /**
- * IdealDetails
+ * MbwayDetails
  */
 
-public class IdealDetails implements PaymentMethodDetails {
-    @SerializedName("issuer")
-    private String issuer = null;
+public class MbwayDetails implements PaymentMethodDetails {
+    @SerializedName("shopperEmail")
+    private String shopperEmail = null;
+
+    @SerializedName("telephoneNumber")
+    private String telephoneNumber = null;
 
     @SerializedName("type")
-    private String type = "ideal";
+    private String type = null;
 
-    public IdealDetails issuer(String issuer) {
-        this.issuer = issuer;
+    public MbwayDetails shopperEmail(String shopperEmail) {
+        this.shopperEmail = shopperEmail;
         return this;
     }
 
     /**
-     * The iDEAL issuer value of the shopper&#x27;s selected bank. Set this to an **id** of an iDEAL issuer to preselect it.
+     * Get shopperEmail
      *
-     * @return issuer
+     * @return shopperEmail
      **/
-    public String getIssuer() {
-        return issuer;
+    public String getShopperEmail() {
+        return shopperEmail;
     }
 
-    public void setIssuer(String issuer) {
-        this.issuer = issuer;
+    public void setShopperEmail(String shopperEmail) {
+        this.shopperEmail = shopperEmail;
     }
 
-    public IdealDetails type(String type) {
+    public MbwayDetails telephoneNumber(String telephoneNumber) {
+        this.telephoneNumber = telephoneNumber;
+        return this;
+    }
+
+    /**
+     * Get telephoneNumber
+     *
+     * @return telephoneNumber
+     **/
+    public String getTelephoneNumber() {
+        return telephoneNumber;
+    }
+
+    public void setTelephoneNumber(String telephoneNumber) {
+        this.telephoneNumber = telephoneNumber;
+    }
+
+    public MbwayDetails type(String type) {
         this.type = type;
         return this;
     }
 
     /**
-     * **ideal**
+     * Get type
      *
      * @return type
      **/
@@ -81,23 +103,25 @@ public class IdealDetails implements PaymentMethodDetails {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        IdealDetails idealDetails = (IdealDetails) o;
-        return Objects.equals(this.issuer, idealDetails.issuer) &&
-                Objects.equals(this.type, idealDetails.type);
+        MbwayDetails mbwayDetails = (MbwayDetails) o;
+        return Objects.equals(this.shopperEmail, mbwayDetails.shopperEmail) &&
+                Objects.equals(this.telephoneNumber, mbwayDetails.telephoneNumber) &&
+                Objects.equals(this.type, mbwayDetails.type);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(issuer, type);
+        return Objects.hash(shopperEmail, telephoneNumber, type);
     }
 
 
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append("class IdealDetails {\n");
+        sb.append("class MbwayDetails {\n");
 
-        sb.append("    issuer: ").append(toIndentedString(issuer)).append("\n");
+        sb.append("    shopperEmail: ").append(toIndentedString(shopperEmail)).append("\n");
+        sb.append("    telephoneNumber: ").append(toIndentedString(telephoneNumber)).append("\n");
         sb.append("    type: ").append(toIndentedString(type)).append("\n");
         sb.append("}");
         return sb.toString();
