+++ src/main/java/com/adyen/model/checkout/details/VisaCheckoutDetails.java
@@ -19,10 +19,9 @@
  * See the LICENSE file for more info.
  */
 
-package com.adyen.model.checkout;
+package com.adyen.model.checkout.details;
 
-import java.util.Objects;
-
+import com.adyen.model.checkout.PaymentMethodDetails;
 import com.google.gson.TypeAdapter;
 import com.google.gson.annotations.JsonAdapter;
 import com.google.gson.annotations.SerializedName;
@@ -30,24 +29,24 @@ import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
 
 import java.io.IOException;
+import java.util.Objects;
 
 /**
- * PaypalDetails
+ * VisaCheckoutDetails
  */
 
-public class PayPalDetails implements PaymentMethodDetails {
-
-    @SerializedName("type")
-    private String type = "paypal";
-
-    @JsonAdapter(SubtypeEnum.Adapter.class)
-    public enum SubtypeEnum {
-        SDK("sdk"),
-        REDIRECT("redirect");
+public class VisaCheckoutDetails implements PaymentMethodDetails {
+    /**
+     * Gets or Sets fundingSource
+     */
+    @JsonAdapter(FundingSourceEnum.Adapter.class)
+    public enum FundingSourceEnum {
+        CREDIT("credit"),
+        DEBIT("debit");
 
         private String value;
 
-        SubtypeEnum(String value) {
+        FundingSourceEnum(String value) {
             this.value = value;
         }
 
@@ -60,8 +59,8 @@ public class PayPalDetails implements PaymentMethodDetails {
             return String.valueOf(value);
         }
 
-        public static SubtypeEnum fromValue(String text) {
-            for (SubtypeEnum b : SubtypeEnum.values()) {
+        public static FundingSourceEnum fromValue(String text) {
+            for (FundingSourceEnum b : FundingSourceEnum.values()) {
                 if (String.valueOf(b.value).equals(text)) {
                     return b;
                 }
@@ -69,30 +68,54 @@ public class PayPalDetails implements PaymentMethodDetails {
             return null;
         }
 
-        public static class Adapter extends TypeAdapter<SubtypeEnum> {
+        public static class Adapter extends TypeAdapter<FundingSourceEnum> {
             @Override
-            public void write(final JsonWriter jsonWriter, final SubtypeEnum enumeration) throws IOException {
+            public void write(final JsonWriter jsonWriter, final FundingSourceEnum enumeration) throws IOException {
                 jsonWriter.value(enumeration.getValue());
             }
 
             @Override
-            public SubtypeEnum read(final JsonReader jsonReader) throws IOException {
+            public FundingSourceEnum read(final JsonReader jsonReader) throws IOException {
                 String value = jsonReader.nextString();
-                return SubtypeEnum.fromValue(String.valueOf(value));
+                return FundingSourceEnum.fromValue(String.valueOf(value));
             }
         }
     }
 
-    @SerializedName("subtype")
-    private SubtypeEnum subtype = SubtypeEnum.REDIRECT;
+    @SerializedName("fundingSource")
+    private FundingSourceEnum fundingSource = null;
 
-    public PayPalDetails type(String type) {
+    @SerializedName("type")
+    private String type = "visacheckout";
+
+    @SerializedName("visaCheckoutCallId")
+    private String visaCheckoutCallId = null;
+
+    public VisaCheckoutDetails fundingSource(FundingSourceEnum fundingSource) {
+        this.fundingSource = fundingSource;
+        return this;
+    }
+
+    /**
+     * Get fundingSource
+     *
+     * @return fundingSource
+     **/
+    public FundingSourceEnum getFundingSource() {
+        return fundingSource;
+    }
+
+    public void setFundingSource(FundingSourceEnum fundingSource) {
+        this.fundingSource = fundingSource;
+    }
+
+    public VisaCheckoutDetails type(String type) {
         this.type = type;
         return this;
     }
 
     /**
-     * **paypal**
+     * **visacheckout**
      *
      * @return type
      **/
@@ -104,17 +127,22 @@ public class PayPalDetails implements PaymentMethodDetails {
         this.type = type;
     }
 
-    public PayPalDetails subtype(SubtypeEnum subtype) {
-        this.subtype = subtype;
+    public VisaCheckoutDetails visaCheckoutCallId(String visaCheckoutCallId) {
+        this.visaCheckoutCallId = visaCheckoutCallId;
         return this;
     }
 
-    public SubtypeEnum getSubtype() {
-        return subtype;
+    /**
+     * Get visaCheckoutCallId
+     *
+     * @return visaCheckoutCallId
+     **/
+    public String getVisaCheckoutCallId() {
+        return visaCheckoutCallId;
     }
 
-    public void setSubtype(SubtypeEnum subtype) {
-        this.subtype = subtype;
+    public void setVisaCheckoutCallId(String visaCheckoutCallId) {
+        this.visaCheckoutCallId = visaCheckoutCallId;
     }
 
 
@@ -126,24 +154,26 @@ public class PayPalDetails implements PaymentMethodDetails {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        PayPalDetails paypalDetails = (PayPalDetails) o;
-        return Objects.equals(this.subtype, paypalDetails.subtype) &&
-                Objects.equals(this.type, paypalDetails.type);
+        VisaCheckoutDetails visaCheckoutDetails = (VisaCheckoutDetails) o;
+        return Objects.equals(this.fundingSource, visaCheckoutDetails.fundingSource) &&
+                Objects.equals(this.type, visaCheckoutDetails.type) &&
+                Objects.equals(this.visaCheckoutCallId, visaCheckoutDetails.visaCheckoutCallId);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(subtype, type);
+        return Objects.hash(fundingSource, type, visaCheckoutCallId);
     }
 
 
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append("class PayPalDetails {\n");
+        sb.append("class VisaCheckoutDetails {\n");
 
-        sb.append("    subtype: ").append(toIndentedString(subtype)).append("\n");
+        sb.append("    fundingSource: ").append(toIndentedString(fundingSource)).append("\n");
         sb.append("    type: ").append(toIndentedString(type)).append("\n");
+        sb.append("    visaCheckoutCallId: ").append(toIndentedString(visaCheckoutCallId)).append("\n");
         sb.append("}");
         return sb.toString();
     }
